Fix narrowing-conversions (part 1 - integers)
warning: narrowing conversion from 'TimeInt' (aka 'unsigned int') to signed type 'int' is implementation-defined [bugprone-narrowing-conversions,cppcoreguidelines-narrowing-conversions]
warning: narrowing conversion from 'unsigned int' to signed type 'int' is implementation-defined [bugprone-narrowing-conversions,cppcoreguidelines-narrowing-conversions]
warning: narrowing conversion from 'size_t' (aka 'unsigned long') to signed type 'int' is implementation-defined [bugprone-narrowing-conversions,cppcoreguidelines-narrowing-conversions]
etc.

and also
warning: narrowing conversion from 'RealD' (aka 'double') to 'unsigned int' [bugprone-narrowing-conversions,cppcoreguidelines-narrowing-conversions]
warning: narrowing conversion from 'double' to 'int' [bugprone-narrowing-conversions,cppcoreguidelines-narrowing-conversions]
warning: narrowing conversion from 'unsigned int' to 'RealF' (aka 'float') [bugprone-narrowing-conversions,cppcoreguidelines-narrowing-conversions]

General approach: change smaller variable to larger type as required by assignment
* mostly int -> unsigned int

Additionally, fixed reverse for-loop in percolate_unsaturated() and infiltrate_water_high()
- counter no longer reaches -1

+++ include/generic.h
@@ -280,7 +280,7 @@ char *Str_ToLower(char *s, char *r);
 int Str_CompareI(char *t, char *s);
 
 char *sw_strtok(
-    char inputStr[], int *startIndex, int *strLen, const char *delim
+    char inputStr[], size_t *startIndex, size_t *strLen, const char *delim
 );
 
 Bool isDelim(char currChar, const char *delim);
@@ -313,9 +313,9 @@ double get_running_sqr(
 
 double final_running_sd(unsigned int n, double ssqr);
 
-double mean(const double values[], int length);
+double mean(const double values[], unsigned int length);
 
-double standardDeviation(double inputArray[], int length);
+double standardDeviation(double inputArray[], unsigned int length);
 
 
 #ifdef __cplusplus
